为assistant相关操作添加了权限认证功能，只运行superAdmin进行操作

+++ projects/app/src/pages/api/core/assistant/del.ts
@@ -2,14 +2,19 @@ import type { NextApiRequest, NextApiResponse } from 'next';
 import { jsonRes } from '@fastgpt/service/common/response';
 import { connectToDatabase } from '@/service/mongo';
 import { MongoAssistant } from '@fastgpt/service/core/assistant/schema';
+import { authUserRole } from '@fastgpt/service/support/permission/auth/user';
+import { TeamMemberRoleEnum } from '@fastgpt/global/support/user/team/constant';
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse<any>) {
   try {
     await connectToDatabase();
     const { assistantId } = req.query as { assistantId: string };
-
     if (!assistantId) {
       throw new Error('参数错误');
+    }
+    const { role } = await authUserRole({ req, authToken: true });
+    if (role !== TeamMemberRoleEnum.superAdmin) {
+      throw new Error('Permission denied');
     }
     await MongoAssistant.deleteOne({
       _id: assistantId
